[EGD-6753] Fix conversion byte shift
In case of `pixPixelSize` and `pixWidth` being both odd, there is at
least one additional byte at the end of each row that should be
discarded.

On the beginning of each row is parsed there is a need to set the
`pixOffset` to be a multiplicity of `pixRowstride` (which is the byte
count per row) instead of relying continues incrementation by the
`pixWidth`.